removed the server build
