+++ custom/views/pages/AboutPage/index.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { NavBar } from 'spee.ch-components';
-import TestTest from 'test/TestTest';
+// import TestTest from 'test/TestTest';
 
 class AboutPage extends React.Component {
   render () {
     return (
         <div>
             <NavBar />
-           <TestTest />
+           {/*<TestTest />*/}
             <div className='row row--padded'>
                 <h3>Welcome to your custom about page!</h3>
                 <p>This is an example of how you can create your own custom pages to replace the default spee.ch pages...</p>
