+++ package.json
@@ -5,11 +5,8 @@
   "main": "index.js",
   "scripts": {
     "configure": "node cli/configure.js",
-    "build": "webpack --config webpack.prod.js",
-    "build-dev": "webpack --config webpack.dev.js",
-    "build-dev-client": "webpack --config webpack.dev.client.js",
-    "build-dev-server": "webpack --config webpack.dev.server.js",
-    "pre-build-dev": "webpack --config webpack.dev.prebuild.js",
+    "build": "webpack --config webpack.config.js",
+    "build-dev": "webpack --config webpack.config.js --watch",
     "start": "node index.js",
     "start-dev": "builder concurrent build-dev start-dev-server",
     "start-dev-server": "nodemon index.js",
